utils/requirements.txt
colorlog~=4.4.0
dataclasses~=0.6
dill~=0.3.2
matplotlib~=3.2.1
numpy~=1.18.4
pillow>=8.2.0
prompt_toolkit~=3.0
PyYAML==5.4
scipy~=1.5.4
pandas~=1.1.5
scikit-learn~=0.23.2
sklearn~=0.0
transformers~=2.1.1
tqdm~=4.46.0
hdbscan~=0.8.15
pytorch_lightning